+++ projects/VoxelRenderer/src/procgen/TextureOctaveNoiseSynthesizer.h
@@ -5,11 +5,11 @@
 class TextureOctaveNoiseSynthesizer : TextureDataSynthesizer
 {
 private:
-    float seed;
+    unsigned int seed;
     int octaves = 3;
     float persistence = 0.5;
 
 public:
-    TextureOctaveNoiseSynthesizer(float seed, int octaves, float persistence);
+    TextureOctaveNoiseSynthesizer(unsigned int seed, int octaves, float persistence);
     void generate(TextureData& textureData) override;
 };
